+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 - Unit tests
 
+### Removed
+-  Removed Slack's default icon and replaced it with nothing. This means if you want an icon with the 
+bot you've to provide a URL for the icon. 
+
 ## [1.0.0] - 2018-12-11
 ### Added
 - Added support for posting report to Github as an issue.